AMBARI-23874. Extract cluster metadata construction to separate class
