+++ ambari-server/src/main/java/org/apache/ambari/server/agent/stomp/MetadataHolder.java
@@ -21,12 +21,12 @@ import java.util.Map;
 
 import org.apache.ambari.server.AmbariException;
 import org.apache.ambari.server.agent.stomp.dto.MetadataCluster;
-import org.apache.ambari.server.controller.AmbariManagementControllerImpl;
 import org.apache.ambari.server.events.ClusterComponentsRepoChangedEvent;
 import org.apache.ambari.server.events.ClusterConfigChangedEvent;
 import org.apache.ambari.server.events.MetadataUpdateEvent;
 import org.apache.ambari.server.events.ServiceInstalledEvent;
 import org.apache.ambari.server.events.publishers.AmbariEventPublisher;
+import org.apache.ambari.server.metadata.ClusterMetadataGenerator;
 import org.apache.ambari.server.state.Cluster;
 import org.apache.ambari.server.state.Clusters;
 import org.apache.commons.collections.MapUtils;
@@ -40,7 +40,7 @@ import com.google.inject.Singleton;
 public class MetadataHolder extends AgentClusterDataHolder<MetadataUpdateEvent> {
 
   @Inject
-  private AmbariManagementControllerImpl ambariManagementController;
+  private ClusterMetadataGenerator metadataGenerator;
 
   @Inject
   private Provider<Clusters> m_clusters;
@@ -52,7 +52,7 @@ public class MetadataHolder extends AgentClusterDataHolder<MetadataUpdateEvent> 
 
   @Override
   public MetadataUpdateEvent getCurrentData() throws AmbariException {
-    return ambariManagementController.getClustersMetadata();
+    return metadataGenerator.getClustersMetadata(m_clusters.get());
   }
 
   @Override
@@ -85,19 +85,19 @@ public class MetadataHolder extends AgentClusterDataHolder<MetadataUpdateEvent> 
   @Subscribe
   public void onConfigsChange(ClusterConfigChangedEvent configChangedEvent) throws AmbariException {
     Cluster cluster = m_clusters.get().getCluster(configChangedEvent.getClusterName());
-    updateData(ambariManagementController.getClusterMetadataOnConfigsUpdate(cluster));
+    updateData(metadataGenerator.getClusterMetadataOnConfigsUpdate(cluster));
 
   }
 
   @Subscribe
   public void onServiceCreate(ServiceInstalledEvent serviceInstalledEvent) throws AmbariException {
     Cluster cluster = m_clusters.get().getCluster(serviceInstalledEvent.getClusterId());
-    updateData(ambariManagementController.getClusterMetadataOnServiceInstall(cluster, serviceInstalledEvent.getServiceName()));
+    updateData(metadataGenerator.getClusterMetadataOnServiceInstall(cluster, serviceInstalledEvent.getServiceName()));
   }
 
   @Subscribe
   public void onClusterComponentsRepoUpdate(ClusterComponentsRepoChangedEvent clusterComponentsRepoChangedEvent) throws AmbariException {
     Cluster cluster = m_clusters.get().getCluster(clusterComponentsRepoChangedEvent.getClusterId());
-    updateData(ambariManagementController.getClusterMetadataOnRepoUpdate(cluster));
+    updateData(metadataGenerator.getClusterMetadataOnRepoUpdate(cluster));
   }
 }
