+++ ambari-server/src/main/java/org/apache/ambari/server/controller/AmbariManagementController.java
@@ -39,7 +39,6 @@ import org.apache.ambari.server.controller.metrics.MetricsCollectorHAManager;
 import org.apache.ambari.server.controller.metrics.timeline.cache.TimelineMetricCacheProvider;
 import org.apache.ambari.server.controller.spi.ResourceAlreadyExistsException;
 import org.apache.ambari.server.events.AmbariEvent;
-import org.apache.ambari.server.events.MetadataUpdateEvent;
 import org.apache.ambari.server.events.publishers.AmbariEventPublisher;
 import org.apache.ambari.server.metadata.RoleCommandOrder;
 import org.apache.ambari.server.orm.entities.ExtensionLinkEntity;
@@ -1010,8 +1009,6 @@ public interface AmbariManagementController {
   MpackResponse getMpack(Long mpackId);
 
   HostRepositories retrieveHostRepositories(Cluster cluster, Host host) throws AmbariException;
-
-  MetadataUpdateEvent getClusterMetadataOnConfigsUpdate(Cluster cluster) throws AmbariException;
 
 }
