+++ ambari-server/src/main/java/org/apache/ambari/server/controller/AmbariManagementControllerImpl.java
@@ -18,42 +18,23 @@
 
 package org.apache.ambari.server.controller;
 
-import static org.apache.ambari.server.agent.ExecutionCommand.KeyNames.AGENT_STACK_RETRY_COUNT;
-import static org.apache.ambari.server.agent.ExecutionCommand.KeyNames.AGENT_STACK_RETRY_ON_UNAVAILABILITY;
 import static org.apache.ambari.server.agent.ExecutionCommand.KeyNames.CLIENTS_TO_UPDATE_CONFIGS;
 import static org.apache.ambari.server.agent.ExecutionCommand.KeyNames.COMMAND_RETRY_ENABLED;
 import static org.apache.ambari.server.agent.ExecutionCommand.KeyNames.COMMAND_TIMEOUT;
 import static org.apache.ambari.server.agent.ExecutionCommand.KeyNames.CUSTOM_FOLDER;
 import static org.apache.ambari.server.agent.ExecutionCommand.KeyNames.DB_DRIVER_FILENAME;
-import static org.apache.ambari.server.agent.ExecutionCommand.KeyNames.DB_NAME;
-import static org.apache.ambari.server.agent.ExecutionCommand.KeyNames.GPL_LICENSE_ACCEPTED;
-import static org.apache.ambari.server.agent.ExecutionCommand.KeyNames.GROUP_LIST;
-import static org.apache.ambari.server.agent.ExecutionCommand.KeyNames.HOOKS_FOLDER;
-import static org.apache.ambari.server.agent.ExecutionCommand.KeyNames.HOST_SYS_PREPPED;
-import static org.apache.ambari.server.agent.ExecutionCommand.KeyNames.JAVA_HOME;
-import static org.apache.ambari.server.agent.ExecutionCommand.KeyNames.JAVA_VERSION;
-import static org.apache.ambari.server.agent.ExecutionCommand.KeyNames.JCE_NAME;
-import static org.apache.ambari.server.agent.ExecutionCommand.KeyNames.JDK_LOCATION;
-import static org.apache.ambari.server.agent.ExecutionCommand.KeyNames.JDK_NAME;
 import static org.apache.ambari.server.agent.ExecutionCommand.KeyNames.MAX_DURATION_OF_RETRIES;
-import static org.apache.ambari.server.agent.ExecutionCommand.KeyNames.MYSQL_JDBC_URL;
-import static org.apache.ambari.server.agent.ExecutionCommand.KeyNames.NOT_MANAGED_HDFS_PATH_LIST;
-import static org.apache.ambari.server.agent.ExecutionCommand.KeyNames.ORACLE_JDBC_URL;
 import static org.apache.ambari.server.agent.ExecutionCommand.KeyNames.PACKAGE_LIST;
 import static org.apache.ambari.server.agent.ExecutionCommand.KeyNames.SCRIPT;
 import static org.apache.ambari.server.agent.ExecutionCommand.KeyNames.SCRIPT_TYPE;
 import static org.apache.ambari.server.agent.ExecutionCommand.KeyNames.SERVICE_PACKAGE_FOLDER;
-import static org.apache.ambari.server.agent.ExecutionCommand.KeyNames.STACK_VERSION;
 import static org.apache.ambari.server.agent.ExecutionCommand.KeyNames.UNLIMITED_KEY_JCE_REQUIRED;
-import static org.apache.ambari.server.agent.ExecutionCommand.KeyNames.USER_GROUPS;
-import static org.apache.ambari.server.agent.ExecutionCommand.KeyNames.USER_LIST;
 import static org.apache.ambari.server.controller.AmbariCustomCommandExecutionHelper.masterToSlaveMappingForDecom;
 
 import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
 import java.lang.reflect.Type;
-import java.net.InetAddress;
 import java.text.MessageFormat;
 import java.util.ArrayList;
 import java.util.Collection;
@@ -69,9 +50,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Set;
-import java.util.SortedMap;
 import java.util.TreeMap;
-import java.util.TreeSet;
 import java.util.UUID;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
@@ -107,12 +86,11 @@ import org.apache.ambari.server.agent.stomp.HostLevelParamsHolder;
 import org.apache.ambari.server.agent.stomp.MetadataHolder;
 import org.apache.ambari.server.agent.stomp.TopologyHolder;
 import org.apache.ambari.server.agent.stomp.dto.HostRepositories;
-import org.apache.ambari.server.agent.stomp.dto.MetadataCluster;
-import org.apache.ambari.server.agent.stomp.dto.MetadataServiceInfo;
 import org.apache.ambari.server.agent.stomp.dto.TopologyCluster;
 import org.apache.ambari.server.agent.stomp.dto.TopologyComponent;
 import org.apache.ambari.server.api.services.AmbariMetaInfo;
 import org.apache.ambari.server.api.services.LoggingService;
+import org.apache.ambari.server.configuration.AmbariConfig;
 import org.apache.ambari.server.configuration.Configuration;
 import org.apache.ambari.server.configuration.Configuration.DatabaseType;
 import org.apache.ambari.server.controller.internal.DeleteHostComponentStatusMetaData;
@@ -130,10 +108,10 @@ import org.apache.ambari.server.controller.spi.Resource;
 import org.apache.ambari.server.controller.spi.ResourceAlreadyExistsException;
 import org.apache.ambari.server.controller.spi.SystemException;
 import org.apache.ambari.server.customactions.ActionDefinition;
-import org.apache.ambari.server.events.MetadataUpdateEvent;
 import org.apache.ambari.server.events.TopologyUpdateEvent;
 import org.apache.ambari.server.events.publishers.AmbariEventPublisher;
 import org.apache.ambari.server.metadata.ActionMetadata;
+import org.apache.ambari.server.metadata.ClusterMetadataGenerator;
 import org.apache.ambari.server.metadata.RoleCommandOrder;
 import org.apache.ambari.server.metadata.RoleCommandOrderProvider;
 import org.apache.ambari.server.orm.dao.ClusterDAO;
@@ -238,7 +216,6 @@ import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang.BooleanUtils;
 import org.apache.commons.lang.StringUtils;
 import org.apache.commons.lang.math.NumberUtils;
-import org.apache.http.client.utils.URIBuilder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -274,9 +251,6 @@ public class AmbariManagementControllerImpl implements AmbariManagementControlle
   private static final String CLUSTER_PHASE_PROPERTY = "phase";
   private static final String CLUSTER_PHASE_INITIAL_INSTALL = "INITIAL_INSTALL";
   private static final String CLUSTER_PHASE_INITIAL_START = "INITIAL_START";
-  private static final String AMBARI_SERVER_HOST = "ambari_server_host";
-  private static final String AMBARI_SERVER_PORT = "ambari_server_port";
-  private static final String AMBARI_SERVER_USE_SSL = "ambari_server_use_ssl";
 
   private static final String BASE_LOG_DIR = "/tmp/ambari";
 
@@ -292,6 +266,7 @@ public class AmbariManagementControllerImpl implements AmbariManagementControlle
   private final Injector injector;
 
   private final Gson gson;
+  private final AmbariConfig ambariConfig;
 
 
   @Inject
@@ -390,28 +365,13 @@ public class AmbariManagementControllerImpl implements AmbariManagementControlle
   @Inject
   private RepositoryVersionHelper repoVersionHelper;
 
+  @Inject
+  private ClusterMetadataGenerator metadataGenerator;
+
   /**
    * The KerberosHelper to help setup for enabling for disabling Kerberos
    */
-  private KerberosHelper kerberosHelper;
-
-  final private String masterHostname;
-  final private Integer masterPort;
-  final private String masterProtocol;
-
-  final private static String JDK_RESOURCE_LOCATION =
-      "/resources/";
-
-  final private static int REPO_URL_CONNECT_TIMEOUT = 3000;
-  final private static int REPO_URL_READ_TIMEOUT = 2000;
-
-  final private String jdkResourceUrl;
-  final private String javaHome;
-  final private String jdkName;
-  final private String jceName;
-  final private String ojdbcUrl;
-  final private String serverDB;
-  final private String mysqljdbcUrl;
+  private final KerberosHelper kerberosHelper;
 
   private boolean ldapSyncInProgress;
 
@@ -436,66 +396,19 @@ public class AmbariManagementControllerImpl implements AmbariManagementControlle
     injector.injectMembers(this);
     gson = injector.getInstance(Gson.class);
     LOG.info("Initializing the AmbariManagementControllerImpl");
-    masterHostname =  InetAddress.getLocalHost().getCanonicalHostName();
     maintenanceStateHelper = injector.getInstance(MaintenanceStateHelper.class);
     kerberosHelper = injector.getInstance(KerberosHelper.class);
     m_metadataHolder = injector.getProvider(MetadataHolder.class);
     m_agentConfigsHolder = injector.getProvider(AgentConfigsHolder.class);
     hostComponentStateDAO = injector.getInstance(HostComponentStateDAO.class);
     serviceComponentDesiredStateDAO = injector.getInstance(ServiceComponentDesiredStateDAO.class);
-    if(configs != null)
-    {
-      if (configs.getApiSSLAuthentication()) {
-        masterProtocol = "https";
-        masterPort = configs.getClientSSLApiPort();
-      } else {
-        masterProtocol = "http";
-        masterPort = configs.getClientApiPort();
-      }
-      jdkResourceUrl = getAmbariServerURI(JDK_RESOURCE_LOCATION);
-      javaHome = configs.getJavaHome();
-      jdkName = configs.getJDKName();
-      jceName = configs.getJCEName();
-      ojdbcUrl = getAmbariServerURI(JDK_RESOURCE_LOCATION + "/" + configs.getOjdbcJarName());
-      mysqljdbcUrl = getAmbariServerURI(JDK_RESOURCE_LOCATION + "/" + configs.getMySQLJarName());
-
-      serverDB = configs.getServerDBName();
-    } else {
-      masterProtocol = null;
-      masterPort = null;
-
-      jdkResourceUrl = null;
-      javaHome = null;
-      jdkName = null;
-      jceName = null;
-      ojdbcUrl = null;
-      mysqljdbcUrl = null;
-      serverDB = null;
-    }
+    ambariConfig = metadataGenerator != null ? metadataGenerator.getAmbariConfig() : new AmbariConfig(configs);
     helper = new AmbariManagementHelper(stackDAO, extensionDAO, linkDAO);
   }
 
   @Override
   public String getAmbariServerURI(String path) {
-    if(masterProtocol==null || masterHostname==null || masterPort==null) {
-      return null;
-    }
-
-    URIBuilder uriBuilder = new URIBuilder();
-    uriBuilder.setScheme(masterProtocol);
-    uriBuilder.setHost(masterHostname);
-    uriBuilder.setPort(masterPort);
-
-    String[] parts = path.split("\\?");
-
-    if (parts.length > 1) {
-      uriBuilder.setPath(parts[0]);
-      uriBuilder.setQuery(parts[1]);
-    } else {
-      uriBuilder.setPath(path);
-    }
-
-    return uriBuilder.toString();
+    return ambariConfig.getAmbariServerURI(path);
   }
 
   @Override
@@ -2090,7 +2003,7 @@ public class AmbariManagementControllerImpl implements AmbariManagementControlle
         }
       }
     }
-    m_metadataHolder.get().updateData(getClusterMetadataOnConfigsUpdate(cluster));
+    m_metadataHolder.get().updateData(metadataGenerator.getClusterMetadataOnConfigsUpdate(cluster));
     m_agentConfigsHolder.get().updateData(cluster.getClusterId(), null);
 
     StackId currentVersion = cluster.getCurrentStackVersion();
@@ -2784,14 +2697,9 @@ public class AmbariManagementControllerImpl implements AmbariManagementControlle
     // Set exec command with 'ClusterSettings' map
     execCmd.setClusterSettings(cluster.getClusterSettingsNameValueMap());
 
-
     // Set exec command with 'StackSettings' map
     // TODO avoid sending in each command, send in "async" metadata
-    Map<String, String> stackSettings = ambariMetaInfo.getStackSettingsNameValueMap(stackId);
-    Map<String, String> stackLevelParams = getMetadataStackLevelParams(cluster, stackId);
-    stackLevelParams.putAll(stackSettings);
-    stackLevelParams.put(STACK_VERSION, stackId.getStackVersion());
-    execCmd.setStackSettings(stackLevelParams);
+    execCmd.setStackSettings(metadataGenerator.getMetadataStackLevelParams(cluster, stackId));
 
     if (databaseType == DatabaseType.ORACLE) {
       hostParams.put(DB_DRIVER_FILENAME, configs.getOjdbcJarName());
@@ -4942,37 +4850,37 @@ public class AmbariManagementControllerImpl implements AmbariManagementControlle
 
   @Override
   public String getJdkResourceUrl() {
-    return jdkResourceUrl;
+    return ambariConfig.getJdkResourceUrl();
   }
 
   @Override
   public String getJavaHome() {
-    return javaHome;
+    return ambariConfig.getJavaHome();
   }
 
   @Override
   public String getJDKName() {
-    return jdkName;
+    return ambariConfig.getJDKName();
   }
 
   @Override
   public String getJCEName() {
-    return jceName;
+    return ambariConfig.getJCEName();
   }
 
   @Override
   public String getServerDB() {
-    return serverDB;
+    return ambariConfig.getServerDB();
   }
 
   @Override
   public String getOjdbcUrl() {
-    return ojdbcUrl;
+    return ambariConfig.getOjdbcUrl();
   }
 
   @Override
   public String getMysqljdbcUrl() {
-    return mysqljdbcUrl;
+    return ambariConfig.getMysqljdbcUrl();
   }
 
   @Override
@@ -5711,7 +5619,7 @@ public class AmbariManagementControllerImpl implements AmbariManagementControlle
         }
       }
 
-      m_metadataHolder.get().updateData(getClusterMetadataOnConfigsUpdate(cluster));
+      m_metadataHolder.get().updateData(metadataGenerator.getClusterMetadataOnConfigsUpdate(cluster));
       m_agentConfigsHolder.get().updateData(cluster.getClusterId(), null);
 
       if (request.getVersion() != null) {
@@ -5738,87 +5646,6 @@ public class AmbariManagementControllerImpl implements AmbariManagementControlle
       serviceConfigVersionResponses.add(serviceConfigVersionResponse);
     }
     return serviceConfigVersionResponses;
-  }
-
-  /**
-   * Collects metadata info about clusters for agent.
-   * @return metadata info about clusters
-   * @throws AmbariException
-   */
-  public MetadataUpdateEvent getClustersMetadata() throws AmbariException {
-    TreeMap<String, MetadataCluster> metadataClusters = new TreeMap<>();
-
-    for (Cluster cl : clusters.getClusters().values()) {
-      StackId stackId = cl.getDesiredStackVersion();
-
-      SecurityType securityType = cl.getSecurityType();
-
-      MetadataCluster metadataCluster = new MetadataCluster(securityType,
-          getMetadataServiceLevelParams(cl),
-          getMetadataClusterLevelParams(cl));
-      metadataClusters.put(Long.toString(cl.getClusterId()), metadataCluster);
-    }
-
-    MetadataUpdateEvent metadataUpdateEvent = new MetadataUpdateEvent(metadataClusters,
-        getMetadataAmbariLevelParams());
-    return metadataUpdateEvent;
-  }
-
-  public MetadataUpdateEvent getClusterMetadata(Cluster cl) throws AmbariException {
-    TreeMap<String, MetadataCluster> metadataClusters = new TreeMap<>();
-    StackId stackId = cl.getDesiredStackVersion();
-
-    SecurityType securityType = cl.getSecurityType();
-
-    MetadataCluster metadataCluster = new MetadataCluster(securityType,
-        getMetadataServiceLevelParams(cl),
-        getMetadataClusterLevelParams(cl));
-    metadataClusters.put(Long.toString(cl.getClusterId()), metadataCluster);
-
-    MetadataUpdateEvent metadataUpdateEvent = new MetadataUpdateEvent(metadataClusters,
-        null);
-    return metadataUpdateEvent;
-  }
-
-  @Override
-  public MetadataUpdateEvent getClusterMetadataOnConfigsUpdate(Cluster cl) throws AmbariException {
-    TreeMap<String, MetadataCluster> metadataClusters = new TreeMap<>();
-    StackId stackId = cl.getDesiredStackVersion();
-
-    MetadataCluster metadataCluster = new MetadataCluster(null,
-        new TreeMap<>(),
-        getMetadataClusterLevelParams(cl));
-    metadataClusters.put(Long.toString(cl.getClusterId()), metadataCluster);
-
-    MetadataUpdateEvent metadataUpdateEvent = new MetadataUpdateEvent(metadataClusters,
-        null);
-    return metadataUpdateEvent;
-  }
-
-  public MetadataUpdateEvent getClusterMetadataOnRepoUpdate(Cluster cl) throws AmbariException {
-    TreeMap<String, MetadataCluster> metadataClusters = new TreeMap<>();
-
-    MetadataCluster metadataCluster = new MetadataCluster(null,
-        getMetadataServiceLevelParams(cl),
-        new TreeMap<>());
-    metadataClusters.put(Long.toString(cl.getClusterId()), metadataCluster);
-
-    MetadataUpdateEvent metadataUpdateEvent = new MetadataUpdateEvent(metadataClusters,
-        null);
-    return metadataUpdateEvent;
-  }
-
-  public MetadataUpdateEvent getClusterMetadataOnServiceInstall(Cluster cl, String serviceName) throws AmbariException {
-    TreeMap<String, MetadataCluster> metadataClusters = new TreeMap<>();
-
-    MetadataCluster metadataCluster = new MetadataCluster(null,
-        getMetadataServiceLevelParams(cl.getService(serviceName)),
-        new TreeMap<>());
-    metadataClusters.put(Long.toString(cl.getClusterId()), metadataCluster);
-
-    MetadataUpdateEvent metadataUpdateEvent = new MetadataUpdateEvent(metadataClusters,
-        null);
-    return metadataUpdateEvent;
   }
 
   private String getClientsToUpdateConfigs(ComponentInfo componentInfo) {
@@ -5901,104 +5728,6 @@ public class AmbariManagementControllerImpl implements AmbariManagementControlle
     commandParams.put(SCRIPT_TYPE, script.getScriptType().toString());
 
     return commandParams;
-  }
-
-  private SortedMap<String, String> getMetadataStackLevelParams(Cluster cluster, StackId stackId) throws AmbariException {
-    SortedMap<String, String> stackLevelParams = new TreeMap<>();
-
-    Map<String, DesiredConfig> clusterDesiredConfigs = cluster.getDesiredConfigs();
-    Set<PropertyInfo> stackProperties = ambariMetaInfo.getStackProperties(stackId.getStackName(), stackId.getStackVersion());
-    Map<String, ServiceInfo> servicesMap = ambariMetaInfo.getServices(stackId.getStackName(), stackId.getStackVersion());
-    Set<PropertyInfo> clusterProperties = ambariMetaInfo.getClusterProperties();
-
-    Map<PropertyInfo, String> users = configHelper.getPropertiesWithPropertyType(PropertyType.USER, cluster, clusterDesiredConfigs, servicesMap, stackProperties, clusterProperties);
-    Set<String> userSet = new TreeSet<>(users.values());
-    String userList = gson.toJson(userSet);
-    stackLevelParams.put(USER_LIST, userList);
-
-    Map<PropertyInfo, String> groups = configHelper.getPropertiesWithPropertyType(PropertyType.GROUP, cluster, clusterDesiredConfigs, servicesMap, stackProperties, clusterProperties);
-    Set<String> groupSet = new TreeSet<>(groups.values());
-    String groupList = gson.toJson(groupSet);
-    stackLevelParams.put(GROUP_LIST, groupList);
-
-    Map<String, Set<String>> userGroupsMap = configHelper.createUserGroupsMap(users, groups);
-    String userGroups = gson.toJson(userGroupsMap);
-    stackLevelParams.put(USER_GROUPS, userGroups);
-
-    Map<PropertyInfo, String> notManagedHdfsPathMap = configHelper.getPropertiesWithPropertyType(PropertyType.NOT_MANAGED_HDFS_PATH, cluster, clusterDesiredConfigs, servicesMap, stackProperties, clusterProperties);
-    Set<String> notManagedHdfsPathSet = configHelper.filterInvalidPropertyValues(notManagedHdfsPathMap, NOT_MANAGED_HDFS_PATH_LIST);
-    String notManagedHdfsPathList = gson.toJson(notManagedHdfsPathSet);
-    stackLevelParams.put(NOT_MANAGED_HDFS_PATH_LIST, notManagedHdfsPathList);
-
-    return stackLevelParams;
-  }
-
-  private TreeMap<String, String> getMetadataClusterLevelParams(Cluster cluster) {
-    TreeMap<String, String> clusterLevelParams = new TreeMap<>();
-    clusterLevelParams.put(HOOKS_FOLDER, configs.getProperty(Configuration.HOOKS_FOLDER));
-    return clusterLevelParams;
-  }
-
-  public TreeMap<String, MetadataServiceInfo> getMetadataServiceLevelParams(Cluster cluster) throws AmbariException {
-    TreeMap<String, MetadataServiceInfo> serviceLevelParams = new TreeMap<>();
-    for (Map.Entry<String, Service> serviceEntry : cluster.getServices().entrySet()) {
-      Service service = serviceEntry.getValue();
-      serviceLevelParams.putAll(getMetadataServiceLevelParams(service));
-    }
-    return serviceLevelParams;
-  }
-
-  public TreeMap<String, MetadataServiceInfo> getMetadataServiceLevelParams(Service service) throws AmbariException {
-    TreeMap<String, MetadataServiceInfo> serviceLevelParams = new TreeMap<>();
-
-    StackId serviceStackId = service.getStackId();
-
-    ServiceInfo serviceInfo = ambariMetaInfo.getService(serviceStackId.getStackName(),
-        serviceStackId.getStackVersion(), service.getName());
-    Long statusCommandTimeout = null;
-    if (serviceInfo.getCommandScript() != null) {
-      statusCommandTimeout = new Long(
-          ambariCustomCommandExecutionHelper.getStatusCommandTimeout(serviceInfo));
-    }
-
-    String servicePackageFolder = serviceInfo.getServicePackageFolder();
-
-    serviceLevelParams.put(serviceInfo.getName(), new MetadataServiceInfo(serviceInfo.getVersion(),
-        serviceInfo.isCredentialStoreEnabled(), statusCommandTimeout, servicePackageFolder));
-
-    return serviceLevelParams;
-  }
-
-  public TreeMap<String, String> getMetadataAmbariLevelParams() throws AmbariException {
-    TreeMap<String, String> ambariLevelParams = new TreeMap<>();
-    ambariLevelParams.put(JDK_LOCATION, getJdkResourceUrl());
-    ambariLevelParams.put(JAVA_HOME, getJavaHome());
-    ambariLevelParams.put(JAVA_VERSION, String.valueOf(configs.getJavaVersion()));
-    ambariLevelParams.put(JDK_NAME, getJDKName());
-    ambariLevelParams.put(JCE_NAME, getJCEName());
-    ambariLevelParams.put(DB_NAME, getServerDB());
-    ambariLevelParams.put(MYSQL_JDBC_URL, getMysqljdbcUrl());
-    ambariLevelParams.put(ORACLE_JDBC_URL, getOjdbcUrl());
-    ambariLevelParams.put(DB_DRIVER_FILENAME, configs.getMySQLJarName());
-    ambariLevelParams.put(HOST_SYS_PREPPED, configs.areHostsSysPrepped());
-    ambariLevelParams.put(AGENT_STACK_RETRY_ON_UNAVAILABILITY, configs.isAgentStackRetryOnInstallEnabled());
-    ambariLevelParams.put(AGENT_STACK_RETRY_COUNT, configs.getAgentStackRetryOnInstallCount());
-
-    boolean serverUseSsl = configs.getApiSSLAuthentication();
-    int port = serverUseSsl ? configs.getClientSSLApiPort() : configs.getClientApiPort();
-    ambariLevelParams.put(AMBARI_SERVER_HOST, StageUtils.getHostName());
-    ambariLevelParams.put(AMBARI_SERVER_PORT, Integer.toString(port));
-    ambariLevelParams.put(AMBARI_SERVER_USE_SSL, Boolean.toString(serverUseSsl));
-
-    for (Map.Entry<String, String> dbConnectorName : configs.getDatabaseConnectorNames().entrySet()) {
-      ambariLevelParams.put(dbConnectorName.getKey(), dbConnectorName.getValue());
-    }
-    for (Map.Entry<String, String> previousDBConnectorName : configs.getPreviousDatabaseConnectorNames().entrySet()) {
-      ambariLevelParams.put(previousDBConnectorName.getKey(), previousDBConnectorName.getValue());
-    }
-    ambariLevelParams.put(GPL_LICENSE_ACCEPTED, configs.getGplLicenseAccepted().toString());
-
-    return ambariLevelParams;
   }
 
   @Override
